big fix for gimp plugins failing to work
The gimp plugins are binaries that are called by the main gimp binary, these binaries try to use the host `ld-*.so` instead of the interpreter that is being shipped. 

To fix the issue we patched a relative interpreter for the plugins and set `LD_LIBRARY_PATH`.

This has its downsides, for example the `AppRun` now has to `cd` into the AppDir before continuing, hopefully a better solution is found.